feat: Add courses field to facultyUpload API
Update the facultyUpload API in facultycontroller.js to include a new field called "courses" in the request body. This allows the user to specify the courses associated with the faculty member being created. The courses are now saved along with the faculty details in the database.

Refactor facultymodels.js to include courses field

Modify the facultymodels.js file to include a new field called "courses" in the facultySchema. This ensures that the "courses" field is properly defined in the database schema and can be used to store the associated courses for each faculty member.

Update CreateFaculty.html to send courses data

Update the CreateFaculty.html file to include an example of sending the "courses" data along with the "name" data in the fetch request. This demonstrates how to send an array of courses as strings to the facultyUpload API endpoint.

Please note that the commit messages have been generated based on the provided code changes and recent commits.